// src/openapi/model/tOOLTYPE.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TOOLTYPE = 'zhU_JIA' | 'beI_JIA' | 'daO_TAO' | 'liaO_JIA' | 'jiA_TOU' | 'toolHolder' | 'sI_GONG' | 'zuaN_TOU' | 'xI_DAO' | 'jiaO_DAO' | 'tanG_DAO' | 'daO_PIAN' | 'daO_BING' | 'peI_JIAN' | 'uN_KNOWN';

export const TOOLTYPE = {
    ZhUJia: 'zhU_JIA' as TOOLTYPE,
    BeIJia: 'beI_JIA' as TOOLTYPE,
    DaOTao: 'daO_TAO' as TOOLTYPE,
    LiaOJia: 'liaO_JIA' as TOOLTYPE,
    JiATou: 'jiA_TOU' as TOOLTYPE,
    ToolHolder: 'toolHolder' as TOOLTYPE,
    SIGong: 'sI_GONG' as TOOLTYPE,
    ZuaNTou: 'zuaN_TOU' as TOOLTYPE,
    XIDao: 'xI_DAO' as TOOLTYPE,
    JiaODao: 'jiaO_DAO' as TOOLTYPE,
    TanGDao: 'tanG_DAO' as TOOLTYPE,
    DaOPian: 'daO_PIAN' as TOOLTYPE,
    DaOBing: 'daO_BING' as TOOLTYPE,
    PeIJian: 'peI_JIAN' as TOOLTYPE,
    UNKnown: 'uN_KNOWN' as TOOLTYPE
};